Remove hostname hack
Changing the hostname was included to work-around a livecd-rootfs issue where the hostname was used to set the ubuntu archive and it was setting it to ftpinternal.  That has been fixed and this can be removed.  Furthermore, at present gethostid() in cosmic libc is segfaulting presently in this environment with this set.