Enable point-in-time apt repo mirror
This patch allows us to pass an apt repo timestamp to builddlivefs.
By specifying a timestamp (seconds since epoc ala `date +%s`)
livecd-rootfs can provide a transparent apt proxy for the repo
using the by-hash data closest to the specified time.  When
used in combination with --image-target this can allow parallel
builds with consistent views of the apt archive.