Enable point-in-time snap store view
This patch allows us to pass a snap cohort key to builddlivefs.
By specifying a snap cohort key (created by `snap-tool cohort-create`
from livecd-rootfs) livecd-rootfs will get a temporary snapshot of the
snap channel map.  When used in combination with --repo-snapshot-stamp
and --image-target we can perform parallel image builds that share
a consistent view of the snap channel map and the apt repo.