Allow environment vars to cofigure build. Fix bug with help text
The user config feature was overly complicated. Dropped it in favor of
letting users specify environment variables to override behavior.

The help message dynamically shows the current values of different
build options. Previously, as soon as the --help flag was processed, we
would print the message and exit. Thiis could happen before we parsed
all the flags, thus the text could be wrong.