bartender/azure: do not wait when deleting group
By default, "az group delete" is a synchronous operation. The CLI
command waits for the group to be fully deleted before returning. This
was causing bartender to hang for a few minutes while cleaning Azure
resources.
This patch passes the "--no-wait" flag to make the operation
asynchronous and allow bartender to return earlier.