Azure: generate a new SSH keypair on each run
The previous way of dealing with SSH authentication was broken, as it
expected the user to have a specific SSH config that would default to
the right keypair.
Instead of asking the user to specify where/what is the SSH key they
want to use, we now automatically generate a new keypair for each run.
The location of the keypair is printed out to let the user know where to
find it in case they want to SSH manually into the build machine (run
with --no-cleanup).