(feat): az wait and proper checks
moved to using az <operation> wait commands instead of until loops. Was
experiencing the occasional odd issue with ip_addr not getting set
because the initial request didn't contain a PublicIP. Adding in the
wait for the VM to be running and checking for the IP ensures that the
next step, which is SSH connectivity, is more likely to succeed.

also adds azure specific trap. Unsure way, but the bartender wrapper
level trap cleanup was not operating on azure. Adding at the provider
level ensured that cleanup did occur during an issue.