fix: sbom-tools match flag convention and default
--hook-extras-sbom-tools-dir was different than all the other *-dir
flags, in that it didn't indicate a local checkout/directory to nest
into bartender, but instead indicated a subdirectory of an sbom-tool to
nest. This made usage unclear. Separated sbom-tools-subdir and
sbom-tools-dir into separate params, following the same convention as
other *-dir params.

Further the description for the -dir mentioned it was useful. It's not
useful, it's a strict requirement to pass when building Canonical Public
Cloud projects that create an SBOM. this is due to the code that
generates the sbom, at this time, referencing the __init__.py in the
cpc-sbom tool. The project is a separate src Python style, so the
__init__.py lives in a nested directory. To run CPC builds, one must
pass in the var, with a specific string. Since it is the default for
using the currently supported sbom-tool, I've made it the default to the
value. This value is unused if no other sbom-tools values are provided.