Pick AWS_PROFILE from environment
AWS_PROFILE is currently forcibly set to default. It can be overwritten
by the --aws-profile flag. But that requires other scripts that use this
script to support passing --aws-profile. Instead of modyfing these other
scripts, pick the AWS_PROFILE from environment and set it to default
only when it's empty or unset.